requirements.txt
tensorflow
opencv-python
numpy
pillow
pygame
scikit-learn
matplotlib
seaborn